add walletService
