+++ src/background/services/index.ts
@@ -2,6 +2,7 @@ export { OpenPaymentsService } from './openPayments';
 export { StorageService } from './storage';
 export { MonetizationService } from './monetization';
 export { OutgoingPaymentGrantService } from './outgoingPaymentGrant';
+export { WalletService } from './wallet';
 export { Background } from './background';
 export { TabEvents } from './tabEvents';
 export { TabState } from './tabState';
