+++ src/background/services/openPayments.ts
@@ -1,50 +1,13 @@
-// cSpell:ignore keyid
-import type { AmountValue } from 'shared/types';
 import {
   type AuthenticatedClient,
   createAuthenticatedClient,
   OpenPaymentsClientError,
 } from '@interledger/open-payments/dist/client';
-import {
-  type IncomingPayment,
-  type OutgoingPaymentWithSpentAmounts as OutgoingPayment,
-  type WalletAddress,
-} from '@interledger/open-payments/dist/types';
 import * as ed from '@noble/ed25519';
 import { type Request } from 'http-message-signatures';
 import { signMessage } from 'http-message-signatures/lib/httpbis';
 import { createContentDigestHeader } from 'httpbis-digest-headers';
-import { getExchangeRates, getRateOfPay } from '@/background/utils';
-import { KeyAutoAddService } from './keyAutoAdd';
-import { exportJWK, generateEd25519KeyPair } from '@/shared/crypto';
-import { bytesToHex } from '@noble/hashes/utils';
-import {
-  ErrorWithKey,
-  errorWithKeyToJSON,
-  getWalletInformation,
-  isErrorWithKey,
-  sleep,
-  type ErrorWithKeyLike,
-} from '@/shared/helpers';
-import type {
-  AddFundsPayload,
-  ConnectWalletPayload,
-  ReconnectWalletPayload,
-  UpdateBudgetPayload,
-} from '@/shared/messages';
-import {
-  DEFAULT_RATE_OF_PAY,
-  MAX_RATE_OF_PAY,
-  MIN_RATE_OF_PAY,
-  OUTGOING_PAYMENT_POLLING_INTERVAL,
-  OUTGOING_PAYMENT_POLLING_INITIAL_DELAY,
-} from '../config';
 import type { Cradle } from '@/background/container';
-import {
-  InteractionIntent,
-  OutgoingPaymentGrantService,
-} from './outgoingPaymentGrant';
-import { APP_URL } from '@/background/constants';
 
 interface KeyInformation {
   privateKey: string;
@@ -72,54 +35,20 @@ interface SignOptions {
   privateKey: Uint8Array;
   keyId: string;
 }
-
-interface CreateOutgoingPaymentParams {
-  walletAddress: WalletAddress;
-  incomingPaymentId: IncomingPayment['id'];
-  amount: string;
-}
-
 export class OpenPaymentsService {
   private browser: Cradle['browser'];
   private storage: Cradle['storage'];
-  private grantService: Cradle['grantService'];
-  // private appName: Cradle['appName'];
-  // private browserName: Cradle['browserName'];
-  private t: Cradle['t'];
 
-  constructor({
-    browser,
-    storage,
-    grantService,
-    appName,
-    browserName,
-    t,
-  }: Cradle) {
+  constructor({ browser, storage }: Cradle) {
     Object.assign(this, {
       browser,
       storage,
-      grantService,
-      appName,
-      browserName,
-      t,
     });
 
     void this.initialize();
   }
 
-  public client?: AuthenticatedClient;
-
-  public switchGrant(): OutgoingPaymentGrantService['switchGrant'] {
-    return this.grantService.switchGrant;
-  }
-
-  public isAnyGrantUsable() {
-    return this.grantService.isAnyGrantUsable();
-  }
-
-  public async rotateToken() {
-    return this.grantService.rotateToken(this.client!);
-  }
+  public client: AuthenticatedClient;
 
   private async initialize() {
     const { connected, walletAddress, oneTimeGrant, recurringGrant } =
@@ -278,409 +207,6 @@ export class OpenPaymentsService {
         return initialRequest;
       },
     });
-  }
-
-  async connectWallet(params: ConnectWalletPayload | null) {
-    if (!params) {
-      this.setConnectState(null);
-      return;
-    }
-    const {
-      walletAddressUrl,
-      amount,
-      recurring,
-      autoKeyAdd,
-      autoKeyAddConsent,
-    } = params;
-
-    const walletAddress = await getWalletInformation(walletAddressUrl);
-    const exchangeRates = await getExchangeRates();
-
-    let rateOfPay = DEFAULT_RATE_OF_PAY;
-    let minRateOfPay = MIN_RATE_OF_PAY;
-    let maxRateOfPay = MAX_RATE_OF_PAY;
-
-    if (!exchangeRates.rates[walletAddress.assetCode]) {
-      throw new Error(
-        `Exchange rate for ${walletAddress.assetCode} not found.`,
-      );
-    }
-
-    const exchangeRate = exchangeRates.rates[walletAddress.assetCode];
-    rateOfPay = getRateOfPay({
-      rate: DEFAULT_RATE_OF_PAY,
-      exchangeRate,
-      assetScale: walletAddress.assetScale,
-    });
-    minRateOfPay = getRateOfPay({
-      rate: MIN_RATE_OF_PAY,
-      exchangeRate,
-      assetScale: walletAddress.assetScale,
-    });
-    maxRateOfPay = getRateOfPay({
-      rate: MAX_RATE_OF_PAY,
-      exchangeRate,
-      assetScale: walletAddress.assetScale,
-    });
-
-    await this.initClient(walletAddress.id);
-    this.setConnectState('connecting');
-    const [existingTab] = await this.browser.tabs.query({
-      url: this.browser.runtime.getURL(APP_URL),
-    });
-    try {
-      await this.grantService.createGrant(
-        this.client!,
-        amount,
-        walletAddress,
-        recurring,
-        InteractionIntent.CONNECT,
-        existingTab?.id,
-      );
-    } catch (error) {
-      if (
-        isErrorWithKey(error) &&
-        error.key === 'connectWallet_error_invalidClient' &&
-        autoKeyAdd
-      ) {
-        if (!KeyAutoAddService.supports(walletAddress)) {
-          this.updateConnectStateError(error);
-          throw new ErrorWithKey(
-            'connectWalletKeyService_error_notImplemented',
-          );
-        }
-        if (!autoKeyAddConsent) {
-          this.updateConnectStateError(error);
-          throw new ErrorWithKey('connectWalletKeyService_error_noConsent');
-        }
-
-        // add key to wallet and try again
-        try {
-          const tabId = await this.grantService.addPublicKeyToWallet(
-            walletAddress,
-            existingTab?.id,
-          );
-          this.setConnectState('connecting');
-          await this.grantService.createGrant(
-            this.client!,
-            amount,
-            walletAddress,
-            recurring,
-            InteractionIntent.CONNECT,
-            tabId,
-          );
-        } catch (error) {
-          this.updateConnectStateError(error);
-          throw error;
-        }
-      } else {
-        this.updateConnectStateError(error);
-        throw error;
-      }
-    }
-
-    await this.storage.set({
-      walletAddress,
-      rateOfPay,
-      minRateOfPay,
-      maxRateOfPay,
-      connected: true,
-    });
-    this.setConnectState(null);
-  }
-
-  async addFunds({ amount, recurring }: AddFundsPayload) {
-    const { walletAddress, ...grants } = await this.storage.get([
-      'walletAddress',
-      'oneTimeGrant',
-      'recurringGrant',
-    ]);
-
-    await this.grantService.createGrant(
-      this.client!,
-      amount,
-      walletAddress!,
-      recurring,
-      InteractionIntent.FUNDS,
-    );
-
-    // cancel existing grants of same type, if any
-    if (grants.oneTimeGrant && !recurring) {
-      await this.grantService.cancelGrant(
-        grants.oneTimeGrant.continue,
-        this.client!,
-      );
-    } else if (grants.recurringGrant && recurring) {
-      await this.grantService.cancelGrant(
-        grants.recurringGrant.continue,
-        this.client!,
-      );
-    }
-
-    await this.storage.setState({ out_of_funds: false });
-  }
-
-  async updateBudget({ amount, recurring }: UpdateBudgetPayload) {
-    const { walletAddress, ...existingGrants } = await this.storage.get([
-      'walletAddress',
-      'oneTimeGrant',
-      'recurringGrant',
-    ]);
-
-    await this.grantService.createGrant(
-      this.client!,
-      amount,
-      walletAddress!,
-      recurring,
-      InteractionIntent.UPDATE_BUDGET,
-    );
-
-    // Revoke all existing grants.
-    // Note: Clear storage only if new grant type is not same as previous grant
-    // type (as completeGrant already sets new grant state)
-    if (existingGrants.oneTimeGrant) {
-      await this.grantService.cancelGrant(
-        existingGrants.oneTimeGrant.continue,
-        this.client!,
-      );
-      if (recurring) {
-        this.storage.set({
-          oneTimeGrant: null,
-          oneTimeGrantSpentAmount: '0',
-        });
-      }
-    }
-    if (existingGrants.recurringGrant) {
-      await this.grantService.cancelGrant(
-        existingGrants.recurringGrant.continue,
-        this.client!,
-      );
-      if (!recurring) {
-        this.storage.set({
-          recurringGrant: null,
-          recurringGrantSpentAmount: '0',
-        });
-      }
-    }
-  }
-
-  private setConnectState(status: 'connecting' | null) {
-    const state = status ? { status } : null;
-    this.storage.setPopupTransientState('connect', () => state);
-  }
-  private updateConnectStateError(err: ErrorWithKeyLike | { message: string }) {
-    this.storage.setPopupTransientState('connect', (state) => {
-      if (state?.status === 'error:key') {
-        return state;
-      }
-      return {
-        status: 'error',
-        error: isErrorWithKey(err) ? errorWithKeyToJSON(err) : err.message,
-      };
-    });
-  }
-
-  async disconnectWallet() {
-    const { recurringGrant, oneTimeGrant } = await this.storage.get([
-      'recurringGrant',
-      'oneTimeGrant',
-    ]);
-    if (!recurringGrant && !oneTimeGrant) {
-      return;
-    }
-    if (recurringGrant) {
-      await this.grantService.cancelGrant(
-        recurringGrant.continue,
-        this.client!,
-      );
-      this.grantService.disableRecurringGrant();
-    }
-    if (oneTimeGrant) {
-      await this.grantService.cancelGrant(oneTimeGrant.continue, this.client!);
-      this.grantService.disableOneTimeGrant();
-    }
-    await this.storage.clear();
-  }
-
-  async generateKeys() {
-    if (await this.storage.keyPairExists()) return;
-
-    const { privateKey, publicKey } = await generateEd25519KeyPair();
-    const keyId = crypto.randomUUID();
-    const jwk = exportJWK(publicKey, keyId);
-
-    await this.storage.set({
-      privateKey: bytesToHex(privateKey),
-      publicKey: btoa(JSON.stringify(jwk)),
-      keyId,
-    });
-  }
-
-  async createOutgoingPayment({
-    walletAddress,
-    amount,
-    incomingPaymentId,
-  }: CreateOutgoingPaymentParams): Promise<OutgoingPayment> {
-    const outgoingPayment = (await this.client!.outgoingPayment.create(
-      {
-        accessToken: this.grantService.accessToken(),
-        url: walletAddress.resourceServer,
-      },
-      {
-        incomingPayment: incomingPaymentId,
-        walletAddress: walletAddress.id,
-        debitAmount: {
-          value: amount,
-          assetCode: walletAddress.assetCode,
-          assetScale: walletAddress.assetScale,
-        },
-        metadata: {
-          source: 'Web Monetization',
-        },
-      },
-    )) as OutgoingPayment;
-
-    if (outgoingPayment.grantSpentDebitAmount) {
-      this.storage.updateSpentAmount(
-        this.grantService.grantType(),
-        outgoingPayment.grantSpentDebitAmount.value,
-      );
-    }
-    await this.storage.setState({ out_of_funds: false });
-
-    return outgoingPayment;
-  }
-
-  /** Polls for the completion of an outgoing payment */
-  async *pollOutgoingPayment(
-    outgoingPaymentId: OutgoingPayment['id'],
-    {
-      signal,
-      maxAttempts = 10,
-    }: Partial<{ signal: AbortSignal; maxAttempts: number }> = {},
-  ): AsyncGenerator<OutgoingPayment, OutgoingPayment, void> {
-    let attempt = 0;
-    await sleep(OUTGOING_PAYMENT_POLLING_INITIAL_DELAY);
-    while (++attempt <= maxAttempts) {
-      try {
-        signal?.throwIfAborted();
-        const outgoingPayment = await this.client!.outgoingPayment.get({
-          url: outgoingPaymentId,
-          accessToken: this.grantService.accessToken(),
-        });
-        yield outgoingPayment;
-        if (
-          outgoingPayment.failed &&
-          outgoingPayment.sentAmount.value === '0'
-        ) {
-          throw new ErrorWithKey('pay_error_outgoingPaymentFailed');
-        }
-        if (
-          outgoingPayment.debitAmount.value === outgoingPayment.sentAmount.value
-        ) {
-          return outgoingPayment; // completed
-        }
-        signal?.throwIfAborted();
-        await sleep(OUTGOING_PAYMENT_POLLING_INTERVAL);
-      } catch (error) {
-        if (
-          isTokenExpiredError(error) ||
-          isMissingGrantPermissionsError(error)
-        ) {
-          // TODO: We can remove the token `actions` check once we've proper RS
-          // errors in place. Then we can handle insufficient grant error
-          // separately clearly.
-          const token = await this.grantService.rotateToken(this.client!);
-          const hasReadAccess = token.access_token.access.find(
-            (e) => e.type === 'outgoing-payment' && e.actions.includes('read'),
-          );
-          if (!hasReadAccess) {
-            throw new OpenPaymentsClientError('InsufficientGrant', {
-              description: error.description,
-            });
-          }
-        } else {
-          throw error;
-        }
-      }
-    }
-
-    throw new ErrorWithKey('pay_warn_outgoingPaymentPollingIncomplete');
-  }
-
-  async probeDebitAmount(
-    amount: AmountValue,
-    incomingPayment: IncomingPayment['id'],
-    sender: WalletAddress,
-  ): Promise<void> {
-    await this.client!.quote.create(
-      {
-        url: sender.resourceServer,
-        accessToken: this.grantService.accessToken(),
-      },
-      {
-        method: 'ilp',
-        receiver: incomingPayment,
-        walletAddress: sender.id,
-        debitAmount: {
-          value: amount,
-          assetCode: sender.assetCode,
-          assetScale: sender.assetScale,
-        },
-      },
-    );
-  }
-
-  private async validateReconnect() {
-    try {
-      await this.grantService.rotateToken(this.client!);
-    } catch (error) {
-      if (isInvalidClientError(error)) {
-        const msg = this.t('connectWallet_error_invalidClient');
-        throw new Error(msg, { cause: error });
-      }
-      throw error;
-    }
-    await this.storage.setState({ key_revoked: false });
-  }
-
-  async reconnectWallet({ autoKeyAddConsent }: ReconnectWalletPayload) {
-    if (!autoKeyAddConsent) {
-      await this.validateReconnect();
-      return;
-    }
-
-    const { walletAddress } = await this.storage.get(['walletAddress']);
-    if (!walletAddress) {
-      throw new Error('reconnectWallet_error_walletAddressMissing');
-    }
-    if (!KeyAutoAddService.supports(walletAddress)) {
-      throw new ErrorWithKey('connectWalletKeyService_error_notImplemented');
-    }
-
-    try {
-      this.setConnectState('connecting');
-      await this.validateReconnect();
-    } catch (error) {
-      if (!isInvalidClientError(error?.cause)) {
-        this.updateConnectStateError(error);
-        throw error;
-      }
-
-      try {
-        // add key to wallet and try again
-        this.grantService.addPublicKeyToWalletWithRotateToken(
-          this.client!,
-          walletAddress,
-        );
-        await this.storage.setState({ key_revoked: false });
-      } catch (error) {
-        this.updateConnectStateError(error);
-        throw error;
-      }
-    }
-
-    this.setConnectState(null);
   }
 }
