+++ src/pages/popup/index.tsx
@@ -1,6 +1,6 @@
 import './index.css';
 
-import '@/pages/shared/lib/react-scan'; // uncomment this to check performance or re-renders
+// import '@/pages/shared/lib/react-scan'; // uncomment this to check performance or re-renders
 import React from 'react';
 import ReactDOM from 'react-dom/client';
